:hammer: Move finger and spartan handling into separate files
