+++ src/rogallo/screens/main/spartan.py
@@ -1,0 +1,118 @@
+"""Provides code for handling a Spartan request."""
+
+##############################################################################
+# Sybaritic imports.
+from sybaritic import Client, Response, SpartanURI, SybariticError
+
+##############################################################################
+# Textual imports.
+from textual.screen import Screen
+
+##############################################################################
+# Local imports.
+from ...cache import ContentCache
+from ...document import Document
+from ...messages import OpenLocation
+from ...mime_checks import is_displayable_mime_type
+from ...types import SpartanURINeedingData
+from ..user_input import UserInput
+from .local_messages import OpenDocument, OpenUnsupportedMIMEType
+from .text_decoder import decode_text
+
+
+##############################################################################
+async def _handle_response(
+    response: Response, request: OpenLocation, screen: Screen[None], cache: ContentCache
+) -> None:
+    """Handle a response from a Spartan request.
+
+    Args:
+        response: The response to handle.
+        request: The original request that generated the response.
+        screen: The screen to post messages to.
+        cache: The content cache to use for caching documents.
+    """
+    assert isinstance(request.location, SpartanURI)
+    uri = response.uri or response.requested_uri or request.location
+
+    # Handle any non-successful response.
+    if not response.status.is_success:
+        screen.notify(
+            f"Error loading {uri}:\n\n{response.status.value} {response.status.name}\n{response.meta}",
+            severity="error",
+            title="Request Error",
+        )
+        return
+
+    # Handle a successful response.
+    if is_displayable_mime_type(response.mime_type):
+        screen.post_message(
+            OpenDocument(
+                document=cache.add_document(
+                    Document(
+                        location=uri,
+                        original_location=request.location,
+                        content=await decode_text(response),
+                        mime_type=response.mime_type,
+                    )
+                ),
+                original_request=request,
+            )
+        )
+    else:
+        screen.post_message(OpenUnsupportedMIMEType(uri, response.mime_type))
+
+
+##############################################################################
+async def handle_spartan_request(
+    request: OpenLocation, client: Client, screen: Screen[None], cache: ContentCache
+) -> None:
+    """Handle a Spartan request.
+
+    Args:
+        request: The Spartan request to handle.
+        client: The client to use for the request.
+        screen: The screen to post messages to.
+        cache: The content cache to use for caching documents.
+    """
+
+    uri = request.location
+    assert isinstance(uri, SpartanURI)
+
+    # If a cached copy of the document exists and the request allows it,
+    # use that instead of making a network request.
+    if (
+        not isinstance(uri, SpartanURINeedingData)
+        and request.allow_cached
+        and (cached_document := cache.get_document(uri))
+    ):
+        screen.post_message(
+            OpenDocument(
+                document=cached_document,
+                original_request=request,
+            )
+        )
+        return
+
+    # If we're looking at a Spartan request that needs data. Let's ask
+    # the user for it.
+    attached_data: str | None = None
+    if isinstance(uri, SpartanURINeedingData) and not (
+        attached_data := await screen.app.push_screen_wait(
+            UserInput(uri, prompt="Spartan request requires data")
+        )
+    ):
+        return
+
+    try:
+        async with await client.request(uri, data=attached_data) as response:
+            await _handle_response(response, request, screen, cache)
+    except SybariticError as error:
+        screen.notify(
+            f"Error loading {uri}:\n\n{error}",
+            severity="error",
+            title="Spartan Error",
+        )
+
+
+### spartan.py ends here
